+++ client/src/app/AppRoutes.test.tsx
@@ -3,7 +3,7 @@ import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import { act, render } from "@testing-library/react";
 import { store } from "client/utils/store";
-import { AppRoutes } from "client/app/Routes";
+import { AppRoutes } from "client/app/AppRoutes";
 import * as loadData from "client/utils/loadData";
 
 jest.spyOn(loadData, "loadGames").mockReturnValue(Promise.resolve());
